requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
isort==5.0.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
python-telegram-bot==20.1
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.4
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.14.0
